pretty much a rewrite of the pools functionality and more

+++ web/backend/src/models/node-data.model.js
@@ -72,7 +72,7 @@ module.exports = function (app) {
       type: DataTypes.BIGINT
     },
     fee: {
-      type: DataTypes.BIGINT
+      type: DataTypes.FLOAT
     }
   },
   {
